LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/input/keyboard-snap-interruption-expected.txt


FAIL Interrupting a snap animation with another key press targeting the same node preserves snapping destination assert_less_than: Scroll should not have finished yet expected a number less than 600 but got 600
